Put `LinkProberStateMachineBase` into `LinkProberStateMachineBase.h`
